upd: string parsing now under rocketry.parse
