+++ rocketry/parse/__init__.py
@@ -5,3 +5,5 @@ from .time import parse_time
 from ._condition import add_condition_parser
 
 from .utils import CondParser, ParserError
+
+from . import _setup_cond_parsers
